+++ R/ggseqmtplot.R
@@ -76,6 +76,7 @@
 #' ggseqmtplot(actcal.seq, no.n = TRUE, error.bar = "SE") +
 #'  coord_flip() +
 #'  theme(axis.text.y=element_blank(),
+#'        axis.ticks.y = element_blank(),
 #'        panel.grid.major.y = element_blank(),
 #'        legend.position = "top")
 ggseqmtplot <- function(seqdata,
@@ -160,6 +161,7 @@ ggseqmtplot <- function(seqdata,
     geom_bar(aes(y = .data$Mean), stat="identity",
              color = ifelse(border == TRUE, "black",
                             "transparent")) +
+    scale_y_continuous(expand = expansion(add = 0)) +
     scale_fill_manual(drop = FALSE,
                       values = cpal) +
     labs(x = "", y = ylabspec) +
@@ -206,7 +208,9 @@ ggseqmtplot <- function(seqdata,
 
   ggmtplot <- ggmtplot +
     theme(axis.title.y = element_text(vjust = +3),
-          plot.margin = margin(15, 15, 10, 15))
+          plot.margin = margin(15, 15, 10, 15),
+          axis.line.x = element_line(size = .3),
+          axis.ticks = element_line(size = .3))
 
 
   return(ggmtplot)