+++ NEWS.md
@@ -1,3 +1,20 @@
+# ggseqplot 0.7.2
+
+* fixed sorting issue for y-axis labels in grouped iplots
+* allow for additional facet_wrap arguments to change plot appearance 
+
+# ggseqplot 0.7.1
+
+* Tweaking the theme: adding ticks and x axis line
+
+# ggseqplot 0.7.0 
+
+* Added option to break down dplots by state (`dissect`)
+
+# ggseqplot 0.6.2.9000 <br><small>(development version)</small>
+
+* Set minimum R version because of R's new pipe (#1, @cbrueffer)
+
 # ggseqplot 0.6.2
 
 * Added a `NEWS.md` file to track changes to the package.